189-190 "We know [...], because the lifetimes of the heap objects in the payload are highly correlated [...]" It is not immediately apparent why this is so. Is this related to the observation of 185--186 ("Most of the time ...") ? That is, some property of those server programs (ie they do need to keep the deserialized payload for the all duration of request handling), or some unfortunate consequence of GC?
—-----------------------------------
201 : "It can be done with destinations"
At this point, it is unclear to me why it is not possible to allocate directly in the compact region, i.e., use the implementation of Table 3 where all object are allocated in the compact region.

205 : "Let's see how a simple parser for S-expression ..."
The transition is a bit rough.
I would clarify the link with previous paragraph by explaining that we now assume that payloads are serialized in S-expressions.